examples/text_to_image/requirements.txt
pip==23.1.2
accelerate==0.16.0
datasets==2.13.1
ftfy==6.1.1
Jinja2==3.0.3
tensorboard==2.11.2
torchvision==0.13.1
transformers==4.30.2
